+++ src/main/java/fi/evident/elasticsearch/voikko/analysis/FinnishTokenizerFactory.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2014 Evident Solutions Oy
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public License
+ * as published by the Free Software Foundation; either version 2.1
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the Lesser GNU General Public License
+ * along with this program. If not, see <​http://www.gnu.org/licenses/>.
+ */
+
+package fi.evident.elasticsearch.voikko.analysis;
+
+import org.apache.lucene.analysis.Tokenizer;
+import org.elasticsearch.common.inject.Inject;
+import org.elasticsearch.common.inject.assistedinject.Assisted;
+import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.index.Index;
+import org.elasticsearch.index.analysis.AbstractTokenizerFactory;
+import org.elasticsearch.index.settings.IndexSettings;
+
+import java.io.Reader;
+
+/**
+ * Factory that creates {@link FinnishTokenizer}.
+ */
+public class FinnishTokenizerFactory extends AbstractTokenizerFactory {
+
+    @Inject
+    public FinnishTokenizerFactory(Index index, @IndexSettings Settings indexSettings, @Assisted String name, @Assisted Settings settings) {
+        super(index, indexSettings, name, settings);
+    }
+
+    @Override
+    public Tokenizer create(Reader reader) {
+        return new FinnishTokenizer(version, reader);
+    }
+}